Change async write batch streaming behavior
There are two size thresholds that are checked after a batch update is
added to a buffer. If the first is not exceeded, do nothing. If the
first is exceeded but the second is not, check if there is a pending
request, looking at the completion queue for responses but do not block.
If there is not send the buffer, else do nothing. Finally, if the second
threshold is exceeded, block waiting for the response.