Watch etcd for changes to the cluster
Nodes send a request to etcd to watch the "info" key, and spawn a thread
that repeatedly reads for updates. That way, they are constantly updated
on who is the master of each shard, and if any key is found to belong to
a different server, they respond with an invalid gRPC status.

Member functions of Info that can be called while watching are
guaranteed to be thread-safe using a lock.