Support shard migrations
At first the cluster state is set as INIT, and connected nodes are
assigned new shards (20 by default). When the first client gets
connected to the cluster, it changes its state to RUNNING, and no more
shards are getting added. Instead, some of the existing shards are
assigned to new nodes, and they start migrating.

A node that is importing shards, serves requests for them without taking
into account the migration. For example a get request for a key that
exists but has not yet been imported, will report that the key is
missing, and a put might be done and then be overwritten by an older
value. This will be fixed soon.