Make shards delete themselves when not referenced
Use std::shared_ptr for shards and don't delete them explicitly. That
way we can safely read from a shard after it has been erased from the
map, as long as we manage to reference it in time and the shared_ptr is
in scope.

That way we no longer need to lock the whole shards object to increase
the reference counter for writes in a single shard.