Refactor the information stored at etcd
The list of nodes (NodeInfo) don't have any information about the node's
shards anymore. Instead we have a list of shards (ShardInfo) that keep
track of the shard's master and the nodes that the shard is migrating
from and to, if it is migrating.

Also now a node's id never changes and one can uniquely identify a node
by its id instead of its address (i.e. when removing a node with
crocksctl). This means that removed nodes stay in the cluster info but
they are cleared and only take up one or two bytes when serialized.

Finally the shard to shard master map, is no longer an unordered_map but
a simple vector because it's easy to create it by just iterating in the
list of shards.