Allow clients to wait until the cluster is healthy
They achieve that by watching the info key and checking the availability
of the nodes after each update. Waiting is optional and can be
configured using the crocks::Options struct defined in
<crocks/options.h>.

In recent versions of etcd, canceling the watch and calling WritesDone
isn't enough make the server finish the RPC, so Finish on the client
blocks forever. This was fixed by explicitly canceling the RPC using
context.TryCancel().

The ClusterImpl::{Get,Put...} implementation had duplicated code which
grew even larger with the new changes so it got deduplicated using
std::function and friends.