Prefer std::shared_mutex over std::mutex
std::shared_mutex is a multiple-readers/single-writer mutex and if there
is no one modifying the protected structures, readers take the mutex
without blocking. As a result, when no migrations are taking place, the
codebase behaves as if it were completely lock-free.

It is available since C++17, so if compiling with -std=c++14 or lower,
it falls back to std::mutex.